Refactor tab view components structure
