+++ src/features/tab-view/components/data-view/components/index.ts
@@ -1,0 +1,1 @@
+export * from './reset-button';
