+++ src/features/tab-view/components/data-view/index.ts
@@ -1,0 +1,1 @@
+export * from './data-view';
