+++ src/features/tab-view/components/index.ts
@@ -1,4 +1,3 @@
 export * from './data-view';
 export * from './data-view-info-pane';
 export * from './tab-error-fallback';
-export * from './schema-tab-view';
