+++ src/features/tab-view/views/index.ts
@@ -1,0 +1,3 @@
+export * from './script-tab-view';
+export * from './file-data-source-tab-view';
+export * from './schema-tab-view';
